Updated job/views.py and added UserProfile migration

+++ backend/job/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view , permission_classes
 from .models import Job
 from .serializers import JobSerializer
 from rest_framework import status
@@ -8,6 +8,7 @@ from rest_framework.pagination import PageNumberPagination
 from django.shortcuts import get_object_or_404
 from django.db.models import Avg , Min , Max , Count
 from .filters import JobsFilter
+from rest_framework.permissions import IsAuthenticated
 
 
 # Create your views here.
@@ -37,15 +38,22 @@ def getJob(request , pk):
     return Response(serializer.data)
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data 
     job = Job.objects.create(**data)
     serializer = JobSerializer(job , many =False)
     return Response(serializer.data)
 
 @api_view(['PUT']) 
+@permission_classes([IsAuthenticated])
 def updateJob(request , pk): 
     job = get_object_or_404(Job , id=pk)
+
+    if job.user != request.user: 
+        return Response({ 'message' : 'You cannot update this job'}, status=status.HTTP_403_FORBIDDEN)
+
     if 'title' in request.data:
         job.title = request.data['title']
     if 'description' in request.data:
@@ -75,8 +83,13 @@ def updateJob(request , pk):
     return Response(serializer.data)
 
 @api_view(['DELETE']) 
+@permission_classes([IsAuthenticated])
 def deleteJob(request , pk):
     job = get_object_or_404(Job , id = pk)
+
+    if job.user != request.user: 
+        return Response({ 'message' : 'You cannot update this job'}, status=status.HTTP_403_FORBIDDEN)
+    
     job.delete()
     return Response({'message' : 'Job is deleted Successfully.'}, status=status.HTTP_200_OK)
